Network.AWS.IoT

Description

AWS IoT

AWS IoT provides secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud. You can discover your custom IoT-Data endpoint to communicate with, configure rules for data processing and integration with other services, organize resources associated with each thing (Thing Registry), configure logging, and create and manage policies and credentials to authenticate things.

For more information about how AWS IoT works, see the Developer Guide .

Service Configuration

ioT :: Service

API version 2015-05-28 of the Amazon IoT SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoT.

CertificateConflictException

_CertificateConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Unable to verify the CA certificate used to sign the device certificate you are attempting to register. This is happens when you have registered more than one CA certificate that has the same subject field and public key.

SqlParseException

_SqlParseException :: AsError a => Getting (First ServiceError) a ServiceError

The Rule-SQL expression can't be parsed correctly.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

TransferConflictException

_TransferConflictException :: AsError a => Getting (First ServiceError) a ServiceError

You can't transfer the certificate because authorization policies are still attached.

CertificateStateException

_CertificateStateException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate operation is not allowed.

RegistrationCodeValidationException

_RegistrationCodeValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The registration code is invalid.

MalformedPolicyException

_MalformedPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

The policy documentation is not valid.

DeleteConflictException

_DeleteConflictException :: AsError a => Getting (First ServiceError) a ServiceError

You can't delete the resource because it is attached to one or more resources.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

CertificateValidationException

_CertificateValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate is invalid.

TransferAlreadyCompletedException

_TransferAlreadyCompletedException :: AsError a => Getting (First ServiceError) a ServiceError

You can't revert the certificate transfer because the transfer is already complete.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

VersionsLimitExceededException

_VersionsLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of policy versions exceeds the limit.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

InternalException

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

VersionConflictException

_VersionConflictException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when the version of a thing passed to a command is different than the version specified with the --version parameter.

UnauthorizedException

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform this operation.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of attached entities exceeds the limit.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListPolicies (Paginated)

module Network.AWS.IoT.ListPolicies

CreatePolicy

module Network.AWS.IoT.CreatePolicy

RegisterCertificate

module Network.AWS.IoT.RegisterCertificate

ListThingPrincipals

module Network.AWS.IoT.ListThingPrincipals

ListPrincipalThings (Paginated)

module Network.AWS.IoT.ListPrincipalThings

GetLoggingOptions

module Network.AWS.IoT.GetLoggingOptions

ListThingTypes (Paginated)

module Network.AWS.IoT.ListThingTypes

CreateCertificateFromCSR

module Network.AWS.IoT.CreateCertificateFromCSR

DeleteThing

module Network.AWS.IoT.DeleteThing

UpdateThing

module Network.AWS.IoT.UpdateThing

CreateThingType

module Network.AWS.IoT.CreateThingType

CancelCertificateTransfer

module Network.AWS.IoT.CancelCertificateTransfer

DeletePolicyVersion

module Network.AWS.IoT.DeletePolicyVersion

DisableTopicRule

module Network.AWS.IoT.DisableTopicRule

CreateTopicRule

module Network.AWS.IoT.CreateTopicRule

CreatePolicyVersion

module Network.AWS.IoT.CreatePolicyVersion

ListCACertificates (Paginated)

module Network.AWS.IoT.ListCACertificates

DeleteTopicRule

module Network.AWS.IoT.DeleteTopicRule

ListPrincipalPolicies (Paginated)

module Network.AWS.IoT.ListPrincipalPolicies

DeleteCACertificate

module Network.AWS.IoT.DeleteCACertificate

UpdateCACertificate

module Network.AWS.IoT.UpdateCACertificate

ListTopicRules (Paginated)

module Network.AWS.IoT.ListTopicRules

TransferCertificate

module Network.AWS.IoT.TransferCertificate

GetTopicRule

module Network.AWS.IoT.GetTopicRule

DescribeThing

module Network.AWS.IoT.DescribeThing

DeletePolicy

module Network.AWS.IoT.DeletePolicy

ListCertificates (Paginated)

module Network.AWS.IoT.ListCertificates

GetPolicyVersion

module Network.AWS.IoT.GetPolicyVersion

DeleteCertificate

module Network.AWS.IoT.DeleteCertificate

UpdateCertificate

module Network.AWS.IoT.UpdateCertificate

ListOutgoingCertificates (Paginated)

module Network.AWS.IoT.ListOutgoingCertificates

DescribeCACertificate

module Network.AWS.IoT.DescribeCACertificate

GetRegistrationCode

module Network.AWS.IoT.GetRegistrationCode

DeleteThingType

module Network.AWS.IoT.DeleteThingType

ListCertificatesByCA (Paginated)

module Network.AWS.IoT.ListCertificatesByCA

AttachThingPrincipal

module Network.AWS.IoT.AttachThingPrincipal

ListThings (Paginated)

module Network.AWS.IoT.ListThings

DetachPrincipalPolicy

module Network.AWS.IoT.DetachPrincipalPolicy

DeleteRegistrationCode

module Network.AWS.IoT.DeleteRegistrationCode

CreateThing

module Network.AWS.IoT.CreateThing

DescribeCertificate

module Network.AWS.IoT.DescribeCertificate

ListPolicyPrincipals (Paginated)

module Network.AWS.IoT.ListPolicyPrincipals

ReplaceTopicRule

module Network.AWS.IoT.ReplaceTopicRule

SetDefaultPolicyVersion

module Network.AWS.IoT.SetDefaultPolicyVersion

ListPolicyVersions

module Network.AWS.IoT.ListPolicyVersions

CreateKeysAndCertificate

module Network.AWS.IoT.CreateKeysAndCertificate

EnableTopicRule

module Network.AWS.IoT.EnableTopicRule

AcceptCertificateTransfer

module Network.AWS.IoT.AcceptCertificateTransfer

GetPolicy

module Network.AWS.IoT.GetPolicy

DescribeEndpoint

module Network.AWS.IoT.DescribeEndpoint

RegisterCACertificate

module Network.AWS.IoT.RegisterCACertificate

SetLoggingOptions

module Network.AWS.IoT.SetLoggingOptions

DescribeThingType

module Network.AWS.IoT.DescribeThingType

AttachPrincipalPolicy

module Network.AWS.IoT.AttachPrincipalPolicy

RejectCertificateTransfer

module Network.AWS.IoT.RejectCertificateTransfer

DetachThingPrincipal

module Network.AWS.IoT.DetachThingPrincipal

DeprecateThingType

module Network.AWS.IoT.DeprecateThingType

Types

AutoRegistrationStatus

data AutoRegistrationStatus

Constructors

Disable

Enable

CACertificateStatus

data CACertificateStatus

Constructors

CACSActive

CACSInactive

CannedAccessControlList

data CannedAccessControlList

Constructors

AWSExecRead

AuthenticatedRead

BucketOwnerFullControl

BucketOwnerRead

LogDeliveryWrite

Private

PublicRead

PublicReadWrite

CertificateStatus

data CertificateStatus

Constructors

Active

Inactive

PendingActivation

PendingTransfer

RegisterInactive

Revoked

DynamoKeyType

data DynamoKeyType

Constructors

Number

String

LogLevel

data LogLevel

Constructors

Debug

Disabled

Error'

Info

Warn

MessageFormat

data MessageFormat

Constructors

JSON

Raw

Action

data Action

Describes the actions associated with a rule.

See: action smart constructor.

action :: Action

Creates a value of Action with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aCloudwatchMetric - Capture a CloudWatch metric.

aDynamoDBv2 - Write to a DynamoDB table. This is a new version of the DynamoDB action. It allows you to write each attribute in an MQTT message payload into a separate DynamoDB column.

aCloudwatchAlarm - Change the state of a CloudWatch alarm.

aSns - Publish to an Amazon SNS topic.

aDynamoDB - Write to a DynamoDB table.

aFirehose - Write to an Amazon Kinesis Firehose stream.

aLambda - Invoke a Lambda function.

aSalesforce - Send a message to a Salesforce IoT Cloud Input Stream.

aKinesis - Write data to an Amazon Kinesis stream.

aS3 - Write to an Amazon S3 bucket.

aElasticsearch - Write data to an Amazon Elasticsearch Service domain.

aRepublish - Publish to another MQTT topic.

aSqs - Publish to an Amazon SQS queue.

aCloudwatchMetric :: Lens' Action (Maybe CloudwatchMetricAction)

Capture a CloudWatch metric.

aDynamoDBv2 :: Lens' Action (Maybe DynamoDBv2Action)

Write to a DynamoDB table. This is a new version of the DynamoDB action. It allows you to write each attribute in an MQTT message payload into a separate DynamoDB column.

aCloudwatchAlarm :: Lens' Action (Maybe CloudwatchAlarmAction)

Change the state of a CloudWatch alarm.

aSns :: Lens' Action (Maybe SNSAction)

Publish to an Amazon SNS topic.

aDynamoDB :: Lens' Action (Maybe DynamoDBAction)

Write to a DynamoDB table.

aFirehose :: Lens' Action (Maybe FirehoseAction)

Write to an Amazon Kinesis Firehose stream.

aLambda :: Lens' Action (Maybe LambdaAction)

Invoke a Lambda function.

aSalesforce :: Lens' Action (Maybe SalesforceAction)

Send a message to a Salesforce IoT Cloud Input Stream.

aKinesis :: Lens' Action (Maybe KinesisAction)

Write data to an Amazon Kinesis stream.

aS3 :: Lens' Action (Maybe S3Action)

Write to an Amazon S3 bucket.

aElasticsearch :: Lens' Action (Maybe ElasticsearchAction)

Write data to an Amazon Elasticsearch Service domain.

aRepublish :: Lens' Action (Maybe RepublishAction)

Publish to another MQTT topic.

aSqs :: Lens' Action (Maybe SqsAction)

Publish to an Amazon SQS queue.

AttributePayload

data AttributePayload

The attribute payload.

See: attributePayload smart constructor.

attributePayload :: AttributePayload

Creates a value of AttributePayload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apAttributes - A JSON string containing up to three key-value pair in JSON format. For example: {"attributes":{"string1":"string2"}}

apMerge - Specifies whether the list of attributes provided in the AttributePayload is merged with the attributes stored in the registry, instead of overwriting them. To remove an attribute, call UpdateThing with an empty attribute value.

apAttributes :: Lens' AttributePayload (HashMap Text Text)

A JSON string containing up to three key-value pair in JSON format. For example: {"attributes":{"string1":"string2"}}

apMerge :: Lens' AttributePayload (Maybe Bool)

Specifies whether the list of attributes provided in the AttributePayload is merged with the attributes stored in the registry, instead of overwriting them. To remove an attribute, call UpdateThing with an empty attribute value.

CACertificate

data CACertificate

A CA certificate.

See: cACertificate smart constructor.

cACertificate :: CACertificate

Creates a value of CACertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cacStatus - The status of the CA certificate. The status value REGISTER_INACTIVE is deprecated and should not be used.

cacCertificateARN - The ARN of the CA certificate.

cacCertificateId - The ID of the CA certificate.

cacCreationDate - The date the CA certificate was created.

cacStatus :: Lens' CACertificate (Maybe CACertificateStatus)

The status of the CA certificate. The status value REGISTER_INACTIVE is deprecated and should not be used.

cacCertificateARN :: Lens' CACertificate (Maybe Text)

The ARN of the CA certificate.

cacCertificateId :: Lens' CACertificate (Maybe Text)

The ID of the CA certificate.

cacCreationDate :: Lens' CACertificate (Maybe UTCTime)

The date the CA certificate was created.

CACertificateDescription

data CACertificateDescription

Describes a CA certificate.

See: cACertificateDescription smart constructor.

cACertificateDescription :: CACertificateDescription

Creates a value of CACertificateDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cacdStatus - The status of a CA certificate.

cacdOwnedBy - The owner of the CA certificate.

cacdCertificatePem - The CA certificate data, in PEM format.

cacdCertificateARN - The CA certificate ARN.

cacdCertificateId - The CA certificate ID.

cacdAutoRegistrationStatus - Whether the CA certificate configured for auto registration of device certificates. Valid values are ENABLE and DISABLE

cacdCreationDate - The date the CA certificate was created.

cacdStatus :: Lens' CACertificateDescription (Maybe CACertificateStatus)

The status of a CA certificate.

cacdOwnedBy :: Lens' CACertificateDescription (Maybe Text)

The owner of the CA certificate.

cacdCertificatePem :: Lens' CACertificateDescription (Maybe Text)

The CA certificate data, in PEM format.

cacdCertificateARN :: Lens' CACertificateDescription (Maybe Text)

The CA certificate ARN.

cacdCertificateId :: Lens' CACertificateDescription (Maybe Text)

The CA certificate ID.

cacdAutoRegistrationStatus :: Lens' CACertificateDescription (Maybe AutoRegistrationStatus)

Whether the CA certificate configured for auto registration of device certificates. Valid values are ENABLE and DISABLE

cacdCreationDate :: Lens' CACertificateDescription (Maybe UTCTime)

The date the CA certificate was created.

Certificate

data Certificate

Information about a certificate.

See: certificate smart constructor.

certificate :: Certificate

Creates a value of Certificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cStatus - The status of the certificate. The status value REGISTER_INACTIVE is deprecated and should not be used.

cCertificateARN - The ARN of the certificate.

cCertificateId - The ID of the certificate.

cCreationDate - The date and time the certificate was created.

cStatus :: Lens' Certificate (Maybe CertificateStatus)

The status of the certificate. The status value REGISTER_INACTIVE is deprecated and should not be used.

cCertificateARN :: Lens' Certificate (Maybe Text)

The ARN of the certificate.

cCertificateId :: Lens' Certificate (Maybe Text)

The ID of the certificate.

cCreationDate :: Lens' Certificate (Maybe UTCTime)

The date and time the certificate was created.

CertificateDescription

data CertificateDescription

Describes a certificate.

See: certificateDescription smart constructor.

certificateDescription :: CertificateDescription

Creates a value of CertificateDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdStatus - The status of the certificate.

cdOwnedBy - The ID of the AWS account that owns the certificate.

cdLastModifiedDate - The date and time the certificate was last modified.

cdCaCertificateId - The certificate ID of the CA certificate used to sign this certificate.

cdPreviousOwnedBy - The ID of the AWS account of the previous owner of the certificate.

cdCertificatePem - The certificate data, in PEM format.

cdCertificateARN - The ARN of the certificate.

cdCertificateId - The ID of the certificate.

cdCreationDate - The date and time the certificate was created.

cdTransferData - The transfer data.

cdStatus :: Lens' CertificateDescription (Maybe CertificateStatus)

The status of the certificate.

cdOwnedBy :: Lens' CertificateDescription (Maybe Text)

The ID of the AWS account that owns the certificate.

cdLastModifiedDate :: Lens' CertificateDescription (Maybe UTCTime)

The date and time the certificate was last modified.

cdCaCertificateId :: Lens' CertificateDescription (Maybe Text)

The certificate ID of the CA certificate used to sign this certificate.

cdPreviousOwnedBy :: Lens' CertificateDescription (Maybe Text)

The ID of the AWS account of the previous owner of the certificate.

cdCertificatePem :: Lens' CertificateDescription (Maybe Text)

The certificate data, in PEM format.

cdCertificateARN :: Lens' CertificateDescription (Maybe Text)

The ARN of the certificate.

cdCertificateId :: Lens' CertificateDescription (Maybe Text)

The ID of the certificate.

cdCreationDate :: Lens' CertificateDescription (Maybe UTCTime)

The date and time the certificate was created.

cdTransferData :: Lens' CertificateDescription (Maybe TransferData)

The transfer data.

CloudwatchAlarmAction

data CloudwatchAlarmAction

Describes an action that updates a CloudWatch alarm.

See: cloudwatchAlarmAction smart constructor.

cloudwatchAlarmAction

Arguments

:: Text
caaRoleARN

-> Text
caaAlarmName

-> Text
caaStateReason

-> Text
caaStateValue

-> CloudwatchAlarmAction

Creates a value of CloudwatchAlarmAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caaRoleARN - The IAM role that allows access to the CloudWatch alarm.

caaAlarmName - The CloudWatch alarm name.

caaStateReason - The reason for the alarm change.

caaStateValue - The value of the alarm state. Acceptable values are: OK, ALARM, INSUFFICIENT_DATA.

caaRoleARN :: Lens' CloudwatchAlarmAction Text

The IAM role that allows access to the CloudWatch alarm.

caaAlarmName :: Lens' CloudwatchAlarmAction Text

The CloudWatch alarm name.

caaStateReason :: Lens' CloudwatchAlarmAction Text

The reason for the alarm change.

caaStateValue :: Lens' CloudwatchAlarmAction Text

The value of the alarm state. Acceptable values are: OK, ALARM, INSUFFICIENT_DATA.

CloudwatchMetricAction

data CloudwatchMetricAction

Describes an action that captures a CloudWatch metric.

See: cloudwatchMetricAction smart constructor.

cloudwatchMetricAction

Arguments

:: Text
cmaRoleARN

-> Text
cmaMetricNamespace

-> Text
cmaMetricName

-> Text
cmaMetricValue

-> Text
cmaMetricUnit

-> CloudwatchMetricAction

Creates a value of CloudwatchMetricAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmaMetricTimestamp - An optional Unix timestamp .

cmaRoleARN - The IAM role that allows access to the CloudWatch metric.

cmaMetricNamespace - The CloudWatch metric namespace name.

cmaMetricName - The CloudWatch metric name.

cmaMetricValue - The CloudWatch metric value.

cmaMetricUnit - The metric unit supported by CloudWatch.

cmaMetricTimestamp :: Lens' CloudwatchMetricAction (Maybe Text)

An optional Unix timestamp .

cmaRoleARN :: Lens' CloudwatchMetricAction Text

The IAM role that allows access to the CloudWatch metric.

cmaMetricNamespace :: Lens' CloudwatchMetricAction Text

The CloudWatch metric namespace name.

cmaMetricName :: Lens' CloudwatchMetricAction Text

The CloudWatch metric name.

cmaMetricValue :: Lens' CloudwatchMetricAction Text

The CloudWatch metric value.

cmaMetricUnit :: Lens' CloudwatchMetricAction Text

The metric unit supported by CloudWatch.

DynamoDBAction

data DynamoDBAction

Describes an action to write to a DynamoDB table.

The tableName , hashKeyField , and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression }.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

See: dynamoDBAction smart constructor.

dynamoDBAction

Arguments

:: Text
ddbaTableName

-> Text
ddbaRoleARN

-> Text
ddbaHashKeyField

-> Text
ddbaHashKeyValue

-> DynamoDBAction

Creates a value of DynamoDBAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbaHashKeyType - The hash key type. Valid values are STRING or NUMBER

ddbaOperation - The type of operation to be performed. This follows the substitution template, so it can be > {operation} , but the substitution must result in one of the following: INSERT , UPDATE , or DELETE .

ddbaRangeKeyType - The range key type. Valid values are STRING or NUMBER

ddbaPayloadField - The action payload. This name can be customized.

ddbaRangeKeyField - The range key name.

ddbaRangeKeyValue - The range key value.

ddbaTableName - The name of the DynamoDB table.

ddbaRoleARN - The ARN of the IAM role that grants access to the DynamoDB table.

ddbaHashKeyField - The hash key name.

ddbaHashKeyValue - The hash key value.

ddbaHashKeyType :: Lens' DynamoDBAction (Maybe DynamoKeyType)

The hash key type. Valid values are STRING or NUMBER

ddbaOperation :: Lens' DynamoDBAction (Maybe Text)

The type of operation to be performed. This follows the substitution template, so it can be > {operation} , but the substitution must result in one of the following: INSERT , UPDATE , or DELETE .

ddbaRangeKeyType :: Lens' DynamoDBAction (Maybe DynamoKeyType)

The range key type. Valid values are STRING or NUMBER

ddbaPayloadField :: Lens' DynamoDBAction (Maybe Text)

The action payload. This name can be customized.

ddbaRangeKeyField :: Lens' DynamoDBAction (Maybe Text)

The range key name.

ddbaRangeKeyValue :: Lens' DynamoDBAction (Maybe Text)

The range key value.

ddbaTableName :: Lens' DynamoDBAction Text

The name of the DynamoDB table.

ddbaRoleARN :: Lens' DynamoDBAction Text

The ARN of the IAM role that grants access to the DynamoDB table.

ddbaHashKeyField :: Lens' DynamoDBAction Text

The hash key name.

ddbaHashKeyValue :: Lens' DynamoDBAction Text

The hash key value.

DynamoDBv2Action

data DynamoDBv2Action

Describes an action to write to a DynamoDB table.

This DynamoDB action writes each attribute in the message payload into it's own column in the DynamoDB table.

See: dynamoDBv2Action smart constructor.

dynamoDBv2Action :: DynamoDBv2Action

Creates a value of DynamoDBv2Action with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddaPutItem - Specifies the DynamoDB table to which the message data will be written. For example: { "dynamoDBv2": { "roleArn": "aws:iam:12341251:my-role" "putItem": { "tableName": "my-table" } } } Each attribute in the message payload will be written to a separate column in the DynamoDB database.

ddaRoleARN - The ARN of the IAM role that grants access to the DynamoDB table.

ddaPutItem :: Lens' DynamoDBv2Action (Maybe PutItemInput)

Specifies the DynamoDB table to which the message data will be written. For example: { "dynamoDBv2": { "roleArn": "aws:iam:12341251:my-role" "putItem": { "tableName": "my-table" } } } Each attribute in the message payload will be written to a separate column in the DynamoDB database.

ddaRoleARN :: Lens' DynamoDBv2Action (Maybe Text)

The ARN of the IAM role that grants access to the DynamoDB table.

ElasticsearchAction

data ElasticsearchAction

Describes an action that writes data to an Amazon Elasticsearch Service domain.

See: elasticsearchAction smart constructor.

elasticsearchAction

Arguments

:: Text
eaRoleARN

-> Text
eaEndpoint

-> Text
eaIndex

-> Text
eaType

-> Text
eaId

-> ElasticsearchAction

Creates a value of ElasticsearchAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eaRoleARN - The IAM role ARN that has access to Elasticsearch.

eaEndpoint - The endpoint of your Elasticsearch domain.

eaIndex - The Elasticsearch index where you want to store your data.

eaType - The type of document you are storing.

eaId - The unique identifier for the document you are storing.

eaRoleARN :: Lens' ElasticsearchAction Text

The IAM role ARN that has access to Elasticsearch.

eaEndpoint :: Lens' ElasticsearchAction Text

The endpoint of your Elasticsearch domain.

eaIndex :: Lens' ElasticsearchAction Text

The Elasticsearch index where you want to store your data.

eaType :: Lens' ElasticsearchAction Text

The type of document you are storing.

eaId :: Lens' ElasticsearchAction Text

The unique identifier for the document you are storing.

FirehoseAction

data FirehoseAction

Describes an action that writes data to an Amazon Kinesis Firehose stream.

See: firehoseAction smart constructor.

firehoseAction

Arguments

:: Text
faRoleARN

-> Text
faDeliveryStreamName

-> FirehoseAction

Creates a value of FirehoseAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

faSeparator - A character separator that will be used to separate records written to the Firehose stream. Valid values are: '\n' (newline), '\t' (tab), '\r\n' (Windows newline), ',' (comma).

faRoleARN - The IAM role that grants access to the Amazon Kinesis Firehost stream.

faDeliveryStreamName - The delivery stream name.

faSeparator :: Lens' FirehoseAction (Maybe Text)

A character separator that will be used to separate records written to the Firehose stream. Valid values are: '\n' (newline), '\t' (tab), '\r\n' (Windows newline), ',' (comma).

faRoleARN :: Lens' FirehoseAction Text

The IAM role that grants access to the Amazon Kinesis Firehost stream.

faDeliveryStreamName :: Lens' FirehoseAction Text

The delivery stream name.

KeyPair

data KeyPair

Describes a key pair.

See: keyPair smart constructor.

keyPair :: KeyPair

Creates a value of KeyPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kpPrivateKey - The private key.

kpPublicKey - The public key.

kpPrivateKey :: Lens' KeyPair (Maybe Text)

The private key.

kpPublicKey :: Lens' KeyPair (Maybe Text)

The public key.

KinesisAction

data KinesisAction

Describes an action to write data to an Amazon Kinesis stream.

See: kinesisAction smart constructor.

kinesisAction

Arguments

:: Text
kaRoleARN

-> Text
kaStreamName

-> KinesisAction

Creates a value of KinesisAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kaPartitionKey - The partition key.

kaRoleARN - The ARN of the IAM role that grants access to the Amazon Kinesis stream.

kaStreamName - The name of the Amazon Kinesis stream.

kaPartitionKey :: Lens' KinesisAction (Maybe Text)

The partition key.

kaRoleARN :: Lens' KinesisAction Text

The ARN of the IAM role that grants access to the Amazon Kinesis stream.

kaStreamName :: Lens' KinesisAction Text

The name of the Amazon Kinesis stream.

LambdaAction

data LambdaAction

Describes an action to invoke a Lambda function.

See: lambdaAction smart constructor.

lambdaAction

Arguments

:: Text
laFunctionARN

-> LambdaAction

Creates a value of LambdaAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laFunctionARN - The ARN of the Lambda function.

laFunctionARN :: Lens' LambdaAction Text

The ARN of the Lambda function.

LoggingOptionsPayload

data LoggingOptionsPayload

Describes the logging options payload.

See: loggingOptionsPayload smart constructor.

loggingOptionsPayload

Arguments

:: Text
lopRoleARN

-> LoggingOptionsPayload

Creates a value of LoggingOptionsPayload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lopLogLevel - The logging level.

lopRoleARN - The ARN of the IAM role that grants access.

lopLogLevel :: Lens' LoggingOptionsPayload (Maybe LogLevel)

The logging level.

lopRoleARN :: Lens' LoggingOptionsPayload Text

The ARN of the IAM role that grants access.

OutgoingCertificate

data OutgoingCertificate

A certificate that has been transfered but not yet accepted.

See: outgoingCertificate smart constructor.

outgoingCertificate :: OutgoingCertificate

Creates a value of OutgoingCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ocTransferDate - The date the transfer was initiated.

ocCertificateARN - The certificate ARN.

ocCertificateId - The certificate ID.

ocTransferredTo - The AWS account to which the transfer was made.

ocCreationDate - The certificate creation date.

ocTransferMessage - The transfer message.

ocTransferDate :: Lens' OutgoingCertificate (Maybe UTCTime)

The date the transfer was initiated.

ocCertificateARN :: Lens' OutgoingCertificate (Maybe Text)

The certificate ARN.

ocCertificateId :: Lens' OutgoingCertificate (Maybe Text)

The certificate ID.

ocTransferredTo :: Lens' OutgoingCertificate (Maybe Text)

The AWS account to which the transfer was made.

ocCreationDate :: Lens' OutgoingCertificate (Maybe UTCTime)

The certificate creation date.

ocTransferMessage :: Lens' OutgoingCertificate (Maybe Text)

The transfer message.

Policy

data Policy

Describes an AWS IoT policy.

See: policy smart constructor.

policy :: Policy

Creates a value of Policy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pPolicyName - The policy name.

pPolicyARN - The policy ARN.

pPolicyName :: Lens' Policy (Maybe Text)

The policy name.

pPolicyARN :: Lens' Policy (Maybe Text)

The policy ARN.

PolicyVersion

data PolicyVersion

Describes a policy version.

See: policyVersion smart constructor.

policyVersion :: PolicyVersion

Creates a value of PolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pvVersionId - The policy version ID.

pvCreateDate - The date and time the policy was created.

pvIsDefaultVersion - Specifies whether the policy version is the default.

pvVersionId :: Lens' PolicyVersion (Maybe Text)

The policy version ID.

pvCreateDate :: Lens' PolicyVersion (Maybe UTCTime)

The date and time the policy was created.

pvIsDefaultVersion :: Lens' PolicyVersion (Maybe Bool)

Specifies whether the policy version is the default.

PutItemInput

data PutItemInput

The input for the DynamoActionVS action that specifies the DynamoDB table to which the message data will be written.

See: putItemInput smart constructor.

putItemInput

Arguments

:: Text
piiTableName

-> PutItemInput

Creates a value of PutItemInput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piiTableName - The table where the message data will be written

piiTableName :: Lens' PutItemInput Text

The table where the message data will be written

RepublishAction

data RepublishAction

Describes an action to republish to another topic.

See: republishAction smart constructor.

republishAction

Arguments

:: Text
raRoleARN

-> Text
raTopic

-> RepublishAction

Creates a value of RepublishAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raRoleARN - The ARN of the IAM role that grants access.

raTopic - The name of the MQTT topic.

raRoleARN :: Lens' RepublishAction Text

The ARN of the IAM role that grants access.

raTopic :: Lens' RepublishAction Text

The name of the MQTT topic.

S3Action

data S3Action

Describes an action to write data to an Amazon S3 bucket.

See: s3Action smart constructor.

s3Action

Arguments

:: Text
sRoleARN

-> Text
sBucketName

-> Text
sKey

-> S3Action

Creates a value of S3Action with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sCannedACL - The Amazon S3 canned ACL that controls access to the object identified by the object key. For more information, see S3 canned ACLs .

sRoleARN - The ARN of the IAM role that grants access.

sBucketName - The Amazon S3 bucket.

sKey - The object key.

sCannedACL :: Lens' S3Action (Maybe CannedAccessControlList)

The Amazon S3 canned ACL that controls access to the object identified by the object key. For more information, see S3 canned ACLs .

sRoleARN :: Lens' S3Action Text

The ARN of the IAM role that grants access.

sBucketName :: Lens' S3Action Text

The Amazon S3 bucket.

sKey :: Lens' S3Action Text

The object key.

SNSAction

data SNSAction

Describes an action to publish to an Amazon SNS topic.

See: snsAction smart constructor.

snsAction

Arguments

:: Text
snsaTargetARN

-> Text
snsaRoleARN

-> SNSAction

Creates a value of SNSAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

snsaMessageFormat - The message format of the message to publish. Optional. Accepted values are JSON and RAW. The default value of the attribute is RAW. SNS uses this setting to determine if the payload should be parsed and relevant platform-specific bits of the payload should be extracted. To read more about SNS message formats, see http://docs.aws.amazon.com/sns/latest/dg/json-formats.html refer to their official documentation.

snsaTargetARN - The ARN of the SNS topic.

snsaRoleARN - The ARN of the IAM role that grants access.

snsaMessageFormat :: Lens' SNSAction (Maybe MessageFormat)

The message format of the message to publish. Optional. Accepted values are JSON and RAW. The default value of the attribute is RAW. SNS uses this setting to determine if the payload should be parsed and relevant platform-specific bits of the payload should be extracted. To read more about SNS message formats, see http://docs.aws.amazon.com/sns/latest/dg/json-formats.html refer to their official documentation.

snsaTargetARN :: Lens' SNSAction Text

The ARN of the SNS topic.

snsaRoleARN :: Lens' SNSAction Text

The ARN of the IAM role that grants access.

SalesforceAction

data SalesforceAction

Describes an action to write a message to a Salesforce IoT Cloud Input Stream.

See: salesforceAction smart constructor.

salesforceAction

Arguments

:: Text
saToken

-> Text
saUrl

-> SalesforceAction

Creates a value of SalesforceAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saToken - The token used to authenticate access to the Salesforce IoT Cloud Input Stream. The token is available from the Salesforce IoT Cloud platform after creation of the Input Stream.

saUrl - The URL exposed by the Salesforce IoT Cloud Input Stream. The URL is available from the Salesforce IoT Cloud platform after creation of the Input Stream.

saToken :: Lens' SalesforceAction Text

The token used to authenticate access to the Salesforce IoT Cloud Input Stream. The token is available from the Salesforce IoT Cloud platform after creation of the Input Stream.

saUrl :: Lens' SalesforceAction Text

The URL exposed by the Salesforce IoT Cloud Input Stream. The URL is available from the Salesforce IoT Cloud platform after creation of the Input Stream.

SqsAction

data SqsAction

Describes an action to publish data to an Amazon SQS queue.

See: sqsAction smart constructor.

sqsAction

Arguments

:: Text
saRoleARN

-> Text
saQueueURL

-> SqsAction

Creates a value of SqsAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saUseBase64 - Specifies whether to use Base64 encoding.

saRoleARN - The ARN of the IAM role that grants access.

saQueueURL - The URL of the Amazon SQS queue.

saUseBase64 :: Lens' SqsAction (Maybe Bool)

Specifies whether to use Base64 encoding.

saRoleARN :: Lens' SqsAction Text

The ARN of the IAM role that grants access.

saQueueURL :: Lens' SqsAction Text

The URL of the Amazon SQS queue.

ThingAttribute

data ThingAttribute

The properties of the thing, including thing name, thing type name, and a list of thing attributes.

See: thingAttribute smart constructor.

thingAttribute :: ThingAttribute

Creates a value of ThingAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

taThingTypeName - The name of the thing type, if the thing has been associated with a type.

taAttributes - A list of thing attributes which are name-value pairs.

taVersion - The version of the thing record in the registry.

taThingName - The name of the thing.

taThingTypeName :: Lens' ThingAttribute (Maybe Text)

The name of the thing type, if the thing has been associated with a type.

taAttributes :: Lens' ThingAttribute (HashMap Text Text)

A list of thing attributes which are name-value pairs.

taVersion :: Lens' ThingAttribute (Maybe Integer)

The version of the thing record in the registry.

taThingName :: Lens' ThingAttribute (Maybe Text)

The name of the thing.

ThingTypeDefinition

data ThingTypeDefinition

The definition of the thing type, including thing type name and description.

See: thingTypeDefinition smart constructor.

thingTypeDefinition :: ThingTypeDefinition

Creates a value of ThingTypeDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ttdThingTypeProperties - The ThingTypeProperties for the thing type.

ttdThingTypeName - The name of the thing type.

ttdThingTypeMetadata - The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

ttdThingTypeProperties :: Lens' ThingTypeDefinition (Maybe ThingTypeProperties)

The ThingTypeProperties for the thing type.

ttdThingTypeName :: Lens' ThingTypeDefinition (Maybe Text)

The name of the thing type.

ttdThingTypeMetadata :: Lens' ThingTypeDefinition (Maybe ThingTypeMetadata)

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

ThingTypeMetadata

data ThingTypeMetadata

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when time was deprecated.

See: thingTypeMetadata smart constructor.

thingTypeMetadata :: ThingTypeMetadata

Creates a value of ThingTypeMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ttmDeprecationDate - The date and time when the thing type was deprecated.

ttmCreationDate - The date and time when the thing type was created.

ttmDeprecated - Whether the thing type is deprecated. If true , no new things could be associated with this type.

ttmDeprecationDate :: Lens' ThingTypeMetadata (Maybe UTCTime)

The date and time when the thing type was deprecated.

ttmCreationDate :: Lens' ThingTypeMetadata (Maybe UTCTime)

The date and time when the thing type was created.

ttmDeprecated :: Lens' ThingTypeMetadata (Maybe Bool)

Whether the thing type is deprecated. If true , no new things could be associated with this type.

ThingTypeProperties

data ThingTypeProperties

The ThingTypeProperties contains information about the thing type including: a thing type description, and a list of searchable thing attribute names.

See: thingTypeProperties smart constructor.

thingTypeProperties :: ThingTypeProperties

Creates a value of ThingTypeProperties with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ttpSearchableAttributes - A list of searchable thing attribute names.

ttpThingTypeDescription - The description of the thing type.

ttpSearchableAttributes :: Lens' ThingTypeProperties [Text]

A list of searchable thing attribute names.

ttpThingTypeDescription :: Lens' ThingTypeProperties (Maybe Text)

The description of the thing type.

TopicRule

data TopicRule

Describes a rule.

See: topicRule smart constructor.

topicRule :: TopicRule

Creates a value of TopicRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trCreatedAt - The date and time the rule was created.

trActions - The actions associated with the rule.

trAwsIotSqlVersion - The version of the SQL rules engine to use when evaluating the rule.

trRuleDisabled - Specifies whether the rule is disabled.

trRuleName - The name of the rule.

trSql - The SQL statement used to query the topic. When using a SQL query with multiple lines, be sure to escape the newline characters.

trDescription - The description of the rule.

trCreatedAt :: Lens' TopicRule (Maybe UTCTime)

The date and time the rule was created.

trActions :: Lens' TopicRule [Action]

The actions associated with the rule.

trAwsIotSqlVersion :: Lens' TopicRule (Maybe Text)

The version of the SQL rules engine to use when evaluating the rule.

trRuleDisabled :: Lens' TopicRule (Maybe Bool)

Specifies whether the rule is disabled.

trRuleName :: Lens' TopicRule (Maybe Text)

The name of the rule.

trSql :: Lens' TopicRule (Maybe Text)

The SQL statement used to query the topic. When using a SQL query with multiple lines, be sure to escape the newline characters.

trDescription :: Lens' TopicRule (Maybe Text)

The description of the rule.

TopicRuleListItem

data TopicRuleListItem

Describes a rule.

See: topicRuleListItem smart constructor.

topicRuleListItem :: TopicRuleListItem

Creates a value of TopicRuleListItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trliCreatedAt - The date and time the rule was created.

trliRuleDisabled - Specifies whether the rule is disabled.

trliRuleName - The name of the rule.

trliRuleARN - The rule ARN.

trliTopicPattern - The pattern for the topic names that apply.

trliCreatedAt :: Lens' TopicRuleListItem (Maybe UTCTime)

The date and time the rule was created.

trliRuleDisabled :: Lens' TopicRuleListItem (Maybe Bool)

Specifies whether the rule is disabled.

trliRuleName :: Lens' TopicRuleListItem (Maybe Text)

The name of the rule.

trliRuleARN :: Lens' TopicRuleListItem (Maybe Text)

The rule ARN.

trliTopicPattern :: Lens' TopicRuleListItem (Maybe Text)

The pattern for the topic names that apply.

TopicRulePayload

data TopicRulePayload

Describes a rule.

See: topicRulePayload smart constructor.

topicRulePayload

Arguments

:: Text
trpSql

-> TopicRulePayload

Creates a value of TopicRulePayload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trpAwsIotSqlVersion - The version of the SQL rules engine to use when evaluating the rule.

trpRuleDisabled - Specifies whether the rule is disabled.

trpDescription - The description of the rule.

trpSql - The SQL statement used to query the topic. For more information, see AWS IoT SQL Reference in the AWS IoT Developer Guide .

trpActions - The actions associated with the rule.

trpAwsIotSqlVersion :: Lens' TopicRulePayload (Maybe Text)

The version of the SQL rules engine to use when evaluating the rule.

trpRuleDisabled :: Lens' TopicRulePayload (Maybe Bool)

Specifies whether the rule is disabled.

trpDescription :: Lens' TopicRulePayload (Maybe Text)

The description of the rule.

trpSql :: Lens' TopicRulePayload Text

The SQL statement used to query the topic. For more information, see AWS IoT SQL Reference in the AWS IoT Developer Guide .

trpActions :: Lens' TopicRulePayload [Action]

The actions associated with the rule.

TransferData

data TransferData

Data used to transfer a certificate to an AWS account.

See: transferData smart constructor.

transferData :: TransferData

Creates a value of TransferData with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdTransferDate - The date the transfer took place.

tdAcceptDate - The date the transfer was accepted.

tdTransferMessage - The transfer message.

tdRejectDate - The date the transfer was rejected.

tdRejectReason - The reason why the transfer was rejected.

tdTransferDate :: Lens' TransferData (Maybe UTCTime)

The date the transfer took place.

tdAcceptDate :: Lens' TransferData (Maybe UTCTime)

The date the transfer was accepted.

tdTransferMessage :: Lens' TransferData (Maybe Text)

The transfer message.

tdRejectDate :: Lens' TransferData (Maybe UTCTime)

The date the transfer was rejected.

tdRejectReason :: Lens' TransferData (Maybe Text)

The reason why the transfer was rejected.