Network.AWS.IoT.AttachPrincipalPolicy

Description

Attaches the specified policy to the specified principal (certificate or other credential).

Creating a Request

attachPrincipalPolicy

Arguments

:: Text
appPolicyName

-> Text
appPrincipal

-> AttachPrincipalPolicy

Creates a value of AttachPrincipalPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

appPolicyName - The policy name.

appPrincipal - The principal, which can be a certificate ARN (as returned from the CreateCertificate operation) or an Amazon Cognito ID.

data AttachPrincipalPolicy

The input for the AttachPrincipalPolicy operation.

See: attachPrincipalPolicy smart constructor.

Request Lenses

appPolicyName :: Lens' AttachPrincipalPolicy Text

The policy name.

appPrincipal :: Lens' AttachPrincipalPolicy Text

The principal, which can be a certificate ARN (as returned from the CreateCertificate operation) or an Amazon Cognito ID.

Destructuring the Response

attachPrincipalPolicyResponse :: AttachPrincipalPolicyResponse

Creates a value of AttachPrincipalPolicyResponse with the minimum fields required to make a request.

data AttachPrincipalPolicyResponse

See: attachPrincipalPolicyResponse smart constructor.