Network.AWS.IoT.AttachThingPrincipal

Description

Attaches the specified principal to the specified thing.

Creating a Request

attachThingPrincipal

Arguments

:: Text
atpThingName

-> Text
atpPrincipal

-> AttachThingPrincipal

Creates a value of AttachThingPrincipal with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atpThingName - The name of the thing.

atpPrincipal - The principal, such as a certificate or other credential.

data AttachThingPrincipal

The input for the AttachThingPrincipal operation.

See: attachThingPrincipal smart constructor.

Request Lenses

atpThingName :: Lens' AttachThingPrincipal Text

The name of the thing.

atpPrincipal :: Lens' AttachThingPrincipal Text

The principal, such as a certificate or other credential.

Destructuring the Response

attachThingPrincipalResponse

Arguments

:: Int
atprsResponseStatus

-> AttachThingPrincipalResponse

Creates a value of AttachThingPrincipalResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atprsResponseStatus - -- | The response status code.

data AttachThingPrincipalResponse

The output from the AttachThingPrincipal operation.

See: attachThingPrincipalResponse smart constructor.

Response Lenses

atprsResponseStatus :: Lens' AttachThingPrincipalResponse Int

- | The response status code.