Network.AWS.IoT.CancelCertificateTransfer

Description

Cancels a pending transfer for the specified certificate.

Note Only the transfer source account can use this operation to cancel a transfer. (Transfer destinations can use RejectCertificateTransfer instead.) After transfer, AWS IoT returns the certificate to the source account in the INACTIVE state. After the destination account has accepted the transfer, the transfer cannot be cancelled.

After a certificate transfer is cancelled, the status of the certificate changes from PENDING_TRANSFER to INACTIVE.

Creating a Request

cancelCertificateTransfer

Arguments

:: Text
cctCertificateId

-> CancelCertificateTransfer

Creates a value of CancelCertificateTransfer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cctCertificateId - The ID of the certificate.

data CancelCertificateTransfer

The input for the CancelCertificateTransfer operation.

See: cancelCertificateTransfer smart constructor.

Request Lenses

cctCertificateId :: Lens' CancelCertificateTransfer Text

The ID of the certificate.

Destructuring the Response

cancelCertificateTransferResponse :: CancelCertificateTransferResponse

Creates a value of CancelCertificateTransferResponse with the minimum fields required to make a request.

data CancelCertificateTransferResponse

See: cancelCertificateTransferResponse smart constructor.