Network.AWS.IoT.CreateCertificateFromCSR

Description

Creates an X.509 certificate using the specified certificate signing request.

Note: The CSR must include a public key that is either an RSA key with a length of at least 2048 bits or an ECC key from NIST P-256 or NIST P-384 curves.

Note: Reusing the same certificate signing request (CSR) results in a distinct certificate.

You can create multiple certificates in a batch by creating a directory, copying multiple .csr files into that directory, and then specifying that directory on the command line. The following commands show how to create a batch of certificates given a batch of CSRs.

Assuming a set of CSRs are located inside of the directory my-csr-directory:

On Linux and OS X, the command is:

 ls my-csr-directory/ | xargs -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

This command lists all of the CSRs in my-csr-directory and pipes each CSR file name to the aws iot create-certificate-from-csr AWS CLI command to create a certificate for the corresponding CSR.

The aws iot create-certificate-from-csr part of the command can also be run in parallel to speed up the certificate creation process:

 ls my-csr-directory/ | xargs -P 10 -I {} aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/{}

On Windows PowerShell, the command to create certificates for all CSRs in my-csr-directory is:

ls -Name my-csr-directory | %{aws iot create-certificate-from-csr --certificate-signing-request file://my-csr-directory/$_}

On a Windows command prompt, the command to create certificates for all CSRs in my-csr-directory is:

forfiles /p my-csr-directory /c "cmd /c aws iot create-certificate-from-csr --certificate-signing-request file://@path"

Creating a Request

createCertificateFromCSR

Arguments

:: Text
ccfcsrCertificateSigningRequest

-> CreateCertificateFromCSR

Creates a value of CreateCertificateFromCSR with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccfcsrSetAsActive - Specifies whether the certificate is active.

ccfcsrCertificateSigningRequest - The certificate signing request (CSR).

data CreateCertificateFromCSR

The input for the CreateCertificateFromCsr operation.

See: createCertificateFromCSR smart constructor.

Request Lenses

ccfcsrSetAsActive :: Lens' CreateCertificateFromCSR (Maybe Bool)

Specifies whether the certificate is active.

ccfcsrCertificateSigningRequest :: Lens' CreateCertificateFromCSR Text

The certificate signing request (CSR).

Destructuring the Response

createCertificateFromCSRResponse

Arguments

:: Int
ccfcsrrsResponseStatus

-> CreateCertificateFromCSRResponse

Creates a value of CreateCertificateFromCSRResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccfcsrrsCertificatePem - The certificate data, in PEM format.

ccfcsrrsCertificateARN - The Amazon Resource Name (ARN) of the certificate. You can use the ARN as a principal for policy operations.

ccfcsrrsCertificateId - The ID of the certificate. Certificate management operations only take a certificateId.

ccfcsrrsResponseStatus - -- | The response status code.

data CreateCertificateFromCSRResponse

The output from the CreateCertificateFromCsr operation.

See: createCertificateFromCSRResponse smart constructor.

Response Lenses

ccfcsrrsCertificatePem :: Lens' CreateCertificateFromCSRResponse (Maybe Text)

The certificate data, in PEM format.

ccfcsrrsCertificateARN :: Lens' CreateCertificateFromCSRResponse (Maybe Text)

The Amazon Resource Name (ARN) of the certificate. You can use the ARN as a principal for policy operations.

ccfcsrrsCertificateId :: Lens' CreateCertificateFromCSRResponse (Maybe Text)

The ID of the certificate. Certificate management operations only take a certificateId.

ccfcsrrsResponseStatus :: Lens' CreateCertificateFromCSRResponse Int

- | The response status code.