Network.AWS.IoT.CreateKeysAndCertificate

Description

Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued public key.

Note This is the only time AWS IoT issues the private key for this certificate, so it is important to keep it in a secure location.

Creating a Request

createKeysAndCertificate :: CreateKeysAndCertificate

Creates a value of CreateKeysAndCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckacSetAsActive - Specifies whether the certificate is active.

data CreateKeysAndCertificate

The input for the CreateKeysAndCertificate operation.

See: createKeysAndCertificate smart constructor.

Request Lenses

ckacSetAsActive :: Lens' CreateKeysAndCertificate (Maybe Bool)

Specifies whether the certificate is active.

Destructuring the Response

createKeysAndCertificateResponse

Arguments

:: Int
ckacrsResponseStatus

-> CreateKeysAndCertificateResponse

Creates a value of CreateKeysAndCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckacrsKeyPair - The generated key pair.

ckacrsCertificatePem - The certificate data, in PEM format.

ckacrsCertificateARN - The ARN of the certificate.

ckacrsCertificateId - The ID of the certificate. AWS IoT issues a default subject name for the certificate (for example, AWS IoT Certificate).

ckacrsResponseStatus - -- | The response status code.

data CreateKeysAndCertificateResponse

The output of the CreateKeysAndCertificate operation.

See: createKeysAndCertificateResponse smart constructor.

Response Lenses

ckacrsKeyPair :: Lens' CreateKeysAndCertificateResponse (Maybe KeyPair)

The generated key pair.

ckacrsCertificatePem :: Lens' CreateKeysAndCertificateResponse (Maybe Text)

The certificate data, in PEM format.

ckacrsCertificateARN :: Lens' CreateKeysAndCertificateResponse (Maybe Text)

The ARN of the certificate.

ckacrsCertificateId :: Lens' CreateKeysAndCertificateResponse (Maybe Text)

The ID of the certificate. AWS IoT issues a default subject name for the certificate (for example, AWS IoT Certificate).

ckacrsResponseStatus :: Lens' CreateKeysAndCertificateResponse Int

- | The response status code.