Network.AWS.IoT.CreatePolicyVersion

Description

Creates a new version of the specified AWS IoT policy. To update a policy, create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must use DeletePolicyVersion to delete an existing version before you create a new one.

Optionally, you can set the new version as the policy's default version. The default version is the operative version (that is, the version that is in effect for the certificates to which the policy is attached).

Creating a Request

createPolicyVersion

Arguments

:: Text
cpvPolicyName

-> Text
cpvPolicyDocument

-> CreatePolicyVersion

Creates a value of CreatePolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvSetAsDefault - Specifies whether the policy version is set as the default. When this parameter is true, the new policy version becomes the operative version (that is, the version that is in effect for the certificates to which the policy is attached).

cpvPolicyName - The policy name.

cpvPolicyDocument - The JSON document that describes the policy. Minimum length of 1. Maximum length of 2048, excluding whitespaces

data CreatePolicyVersion

The input for the CreatePolicyVersion operation.

See: createPolicyVersion smart constructor.

Request Lenses

cpvSetAsDefault :: Lens' CreatePolicyVersion (Maybe Bool)

Specifies whether the policy version is set as the default. When this parameter is true, the new policy version becomes the operative version (that is, the version that is in effect for the certificates to which the policy is attached).

cpvPolicyName :: Lens' CreatePolicyVersion Text

The policy name.

cpvPolicyDocument :: Lens' CreatePolicyVersion Text

The JSON document that describes the policy. Minimum length of 1. Maximum length of 2048, excluding whitespaces

Destructuring the Response

createPolicyVersionResponse

Arguments

:: Int
cpvrsResponseStatus

-> CreatePolicyVersionResponse

Creates a value of CreatePolicyVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvrsPolicyDocument - The JSON document that describes the policy.

cpvrsPolicyVersionId - The policy version ID.

cpvrsPolicyARN - The policy ARN.

cpvrsIsDefaultVersion - Specifies whether the policy version is the default.

cpvrsResponseStatus - -- | The response status code.

data CreatePolicyVersionResponse

The output of the CreatePolicyVersion operation.

See: createPolicyVersionResponse smart constructor.

Response Lenses

cpvrsPolicyDocument :: Lens' CreatePolicyVersionResponse (Maybe Text)

The JSON document that describes the policy.

cpvrsPolicyVersionId :: Lens' CreatePolicyVersionResponse (Maybe Text)

The policy version ID.

cpvrsPolicyARN :: Lens' CreatePolicyVersionResponse (Maybe Text)

The policy ARN.

cpvrsIsDefaultVersion :: Lens' CreatePolicyVersionResponse (Maybe Bool)

Specifies whether the policy version is the default.

cpvrsResponseStatus :: Lens' CreatePolicyVersionResponse Int

- | The response status code.