Network.AWS.IoT.CreateThing

Description

Creates a thing record in the thing registry.

Creating a Request

createThing

Arguments

:: Text
ctThingName

-> CreateThing

Creates a value of CreateThing with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctThingTypeName - The name of the thing type associated with the new thing.

ctAttributePayload - The attribute payload, which consists of up to three name/value pairs in a JSON document. For example: {"attributes":{"string1":"string2"}}

ctThingName - The name of the thing to create.

data CreateThing

The input for the CreateThing operation.

See: createThing smart constructor.

Request Lenses

ctThingTypeName :: Lens' CreateThing (Maybe Text)

The name of the thing type associated with the new thing.

ctAttributePayload :: Lens' CreateThing (Maybe AttributePayload)

The attribute payload, which consists of up to three name/value pairs in a JSON document. For example: {"attributes":{"string1":"string2"}}

ctThingName :: Lens' CreateThing Text

The name of the thing to create.

Destructuring the Response

createThingResponse

Arguments

:: Int
ctrsResponseStatus

-> CreateThingResponse

Creates a value of CreateThingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrsThingARN - The ARN of the new thing.

ctrsThingName - The name of the new thing.

ctrsResponseStatus - -- | The response status code.

data CreateThingResponse

The output of the CreateThing operation.

See: createThingResponse smart constructor.

Response Lenses

ctrsThingARN :: Lens' CreateThingResponse (Maybe Text)

The ARN of the new thing.

ctrsThingName :: Lens' CreateThingResponse (Maybe Text)

The name of the new thing.

ctrsResponseStatus :: Lens' CreateThingResponse Int

- | The response status code.