Network.AWS.IoT.CreateTopicRule

Description

Creates a rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.

Creating a Request

createTopicRule

Arguments

:: Text
ctrRuleName

-> TopicRulePayload
ctrTopicRulePayload

-> CreateTopicRule

Creates a value of CreateTopicRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrRuleName - The name of the rule.

ctrTopicRulePayload - The rule payload.

data CreateTopicRule

The input for the CreateTopicRule operation.

See: createTopicRule smart constructor.

Request Lenses

ctrRuleName :: Lens' CreateTopicRule Text

The name of the rule.

ctrTopicRulePayload :: Lens' CreateTopicRule TopicRulePayload

The rule payload.

Destructuring the Response

createTopicRuleResponse :: CreateTopicRuleResponse

Creates a value of CreateTopicRuleResponse with the minimum fields required to make a request.

data CreateTopicRuleResponse

See: createTopicRuleResponse smart constructor.