Network.AWS.IoT.DescribeCACertificate

Description

Describes a registered CA certificate.

Creating a Request

describeCACertificate

Arguments

:: Text
dCertificateId

-> DescribeCACertificate

Creates a value of DescribeCACertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dCertificateId - The CA certificate identifier.

data DescribeCACertificate

The input for the DescribeCACertificate operation.

See: describeCACertificate smart constructor.

Request Lenses

dCertificateId :: Lens' DescribeCACertificate Text

The CA certificate identifier.

Destructuring the Response

describeCACertificateResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeCACertificateResponse

Creates a value of DescribeCACertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsCertificateDescription - The CA certificate description.

desrsResponseStatus - -- | The response status code.

data DescribeCACertificateResponse

The output from the DescribeCACertificate operation.

See: describeCACertificateResponse smart constructor.

Response Lenses

desrsCertificateDescription :: Lens' DescribeCACertificateResponse (Maybe CACertificateDescription)

The CA certificate description.

desrsResponseStatus :: Lens' DescribeCACertificateResponse Int

- | The response status code.