Network.AWS.IoT.DescribeThing

Description

Gets information about the specified thing.

Creating a Request

describeThing

Arguments

:: Text
dThingName

-> DescribeThing

Creates a value of DescribeThing with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dThingName - The name of the thing.

data DescribeThing

The input for the DescribeThing operation.

See: describeThing smart constructor.

Request Lenses

dThingName :: Lens' DescribeThing Text

The name of the thing.

Destructuring the Response

describeThingResponse

Arguments

:: Int
dtrsResponseStatus

-> DescribeThingResponse

Creates a value of DescribeThingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsDefaultClientId - The default client ID.

dtrsThingTypeName - The thing type name.

dtrsAttributes - The thing attributes.

dtrsVersion - The current version of the thing record in the registry.

dtrsThingName - The name of the thing.

dtrsResponseStatus - -- | The response status code.

data DescribeThingResponse

The output from the DescribeThing operation.

See: describeThingResponse smart constructor.

Response Lenses

dtrsDefaultClientId :: Lens' DescribeThingResponse (Maybe Text)

The default client ID.

dtrsThingTypeName :: Lens' DescribeThingResponse (Maybe Text)

The thing type name.

dtrsAttributes :: Lens' DescribeThingResponse (HashMap Text Text)

The thing attributes.

dtrsVersion :: Lens' DescribeThingResponse (Maybe Integer)

The current version of the thing record in the registry.

dtrsThingName :: Lens' DescribeThingResponse (Maybe Text)

The name of the thing.

dtrsResponseStatus :: Lens' DescribeThingResponse Int

- | The response status code.