Network.AWS.IoT.DescribeThingType

Description

Gets information about the specified thing type.

Creating a Request

describeThingType

Arguments

:: Text
dThingTypeName

-> DescribeThingType

Creates a value of DescribeThingType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dThingTypeName - The name of the thing type.

data DescribeThingType

The input for the DescribeThingType operation.

See: describeThingType smart constructor.

Request Lenses

dThingTypeName :: Lens' DescribeThingType Text

The name of the thing type.

Destructuring the Response

describeThingTypeResponse

Arguments

:: Int
drsResponseStatus

-> DescribeThingTypeResponse

Creates a value of DescribeThingTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsThingTypeProperties - The ThingTypeProperties contains information about the thing type including description, and a list of searchable thing attribute names.

drsThingTypeName - The name of the thing type.

drsThingTypeMetadata - The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

drsResponseStatus - -- | The response status code.

data DescribeThingTypeResponse

The output for the DescribeThingType operation.

See: describeThingTypeResponse smart constructor.

Response Lenses

drsThingTypeProperties :: Lens' DescribeThingTypeResponse (Maybe ThingTypeProperties)

The ThingTypeProperties contains information about the thing type including description, and a list of searchable thing attribute names.

drsThingTypeName :: Lens' DescribeThingTypeResponse (Maybe Text)

The name of the thing type.

drsThingTypeMetadata :: Lens' DescribeThingTypeResponse (Maybe ThingTypeMetadata)

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

drsResponseStatus :: Lens' DescribeThingTypeResponse Int

- | The response status code.