Network.AWS.IoT.GetLoggingOptions

Description

Gets the logging options.

Creating a Request

getLoggingOptions :: GetLoggingOptions

Creates a value of GetLoggingOptions with the minimum fields required to make a request.

data GetLoggingOptions

The input for the GetLoggingOptions operation.

See: getLoggingOptions smart constructor.

Destructuring the Response

getLoggingOptionsResponse

Arguments

:: Int
glorsResponseStatus

-> GetLoggingOptionsResponse

Creates a value of GetLoggingOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

glorsLogLevel - The logging level.

glorsRoleARN - The ARN of the IAM role that grants access.

glorsResponseStatus - -- | The response status code.

data GetLoggingOptionsResponse

The output from the GetLoggingOptions operation.

See: getLoggingOptionsResponse smart constructor.

Response Lenses

glorsLogLevel :: Lens' GetLoggingOptionsResponse (Maybe LogLevel)

The logging level.

glorsRoleARN :: Lens' GetLoggingOptionsResponse (Maybe Text)

The ARN of the IAM role that grants access.

glorsResponseStatus :: Lens' GetLoggingOptionsResponse Int

- | The response status code.