Network.AWS.IoT.GetPolicy

Description

Gets information about the specified policy with the policy document of the default version.

Creating a Request

getPolicy

Arguments

:: Text
gpPolicyName

-> GetPolicy

Creates a value of GetPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpPolicyName - The name of the policy.

data GetPolicy

The input for the GetPolicy operation.

See: getPolicy smart constructor.

Request Lenses

gpPolicyName :: Lens' GetPolicy Text

The name of the policy.

Destructuring the Response

getPolicyResponse

Arguments

:: Int
gprsResponseStatus

-> GetPolicyResponse

Creates a value of GetPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gprsPolicyName - The policy name.

gprsPolicyDocument - The JSON document that describes the policy.

gprsDefaultVersionId - The default policy version ID.

gprsPolicyARN - The policy ARN.

gprsResponseStatus - -- | The response status code.

data GetPolicyResponse

The output from the GetPolicy operation.

See: getPolicyResponse smart constructor.

Response Lenses

gprsPolicyName :: Lens' GetPolicyResponse (Maybe Text)

The policy name.

gprsPolicyDocument :: Lens' GetPolicyResponse (Maybe Text)

The JSON document that describes the policy.

gprsDefaultVersionId :: Lens' GetPolicyResponse (Maybe Text)

The default policy version ID.

gprsPolicyARN :: Lens' GetPolicyResponse (Maybe Text)

The policy ARN.

gprsResponseStatus :: Lens' GetPolicyResponse Int

- | The response status code.