Network.AWS.IoT.GetPolicyVersion

Description

Gets information about the specified policy version.

Creating a Request

getPolicyVersion

Arguments

:: Text
gpvPolicyName

-> Text
gpvPolicyVersionId

-> GetPolicyVersion

Creates a value of GetPolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpvPolicyName - The name of the policy.

gpvPolicyVersionId - The policy version ID.

data GetPolicyVersion

The input for the GetPolicyVersion operation.

See: getPolicyVersion smart constructor.

Request Lenses

gpvPolicyName :: Lens' GetPolicyVersion Text

The name of the policy.

gpvPolicyVersionId :: Lens' GetPolicyVersion Text

The policy version ID.

Destructuring the Response

getPolicyVersionResponse

Arguments

:: Int
gpvrsResponseStatus

-> GetPolicyVersionResponse

Creates a value of GetPolicyVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpvrsPolicyName - The policy name.

gpvrsPolicyDocument - The JSON document that describes the policy.

gpvrsPolicyVersionId - The policy version ID.

gpvrsPolicyARN - The policy ARN.

gpvrsIsDefaultVersion - Specifies whether the policy version is the default.

gpvrsResponseStatus - -- | The response status code.

data GetPolicyVersionResponse

The output from the GetPolicyVersion operation.

See: getPolicyVersionResponse smart constructor.

Response Lenses

gpvrsPolicyName :: Lens' GetPolicyVersionResponse (Maybe Text)

The policy name.

gpvrsPolicyDocument :: Lens' GetPolicyVersionResponse (Maybe Text)

The JSON document that describes the policy.

gpvrsPolicyVersionId :: Lens' GetPolicyVersionResponse (Maybe Text)

The policy version ID.

gpvrsPolicyARN :: Lens' GetPolicyVersionResponse (Maybe Text)

The policy ARN.

gpvrsIsDefaultVersion :: Lens' GetPolicyVersionResponse (Maybe Bool)

Specifies whether the policy version is the default.

gpvrsResponseStatus :: Lens' GetPolicyVersionResponse Int

- | The response status code.