Network.AWS.IoT.ListThingTypes

Description

Lists the existing thing types.

This operation returns paginated results.

Creating a Request

listThingTypes :: ListThingTypes

Creates a value of ListThingTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lttThingTypeName - The name of the thing type.

lttNextToken - The token for the next set of results, or null if there are no additional results.

lttMaxResults - The maximum number of results to return in this operation.

data ListThingTypes

The input for the ListThingTypes operation.

See: listThingTypes smart constructor.

Request Lenses

lttThingTypeName :: Lens' ListThingTypes (Maybe Text)

The name of the thing type.

lttNextToken :: Lens' ListThingTypes (Maybe Text)

The token for the next set of results, or null if there are no additional results.

lttMaxResults :: Lens' ListThingTypes (Maybe Natural)

The maximum number of results to return in this operation.

Destructuring the Response

listThingTypesResponse

Arguments

:: Int
lttrsResponseStatus

-> ListThingTypesResponse

Creates a value of ListThingTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lttrsThingTypes - The thing types.

lttrsNextToken - The token for the next set of results, or null if there are no additional results.

lttrsResponseStatus - -- | The response status code.

data ListThingTypesResponse

The output for the ListThingTypes operation.

See: listThingTypesResponse smart constructor.

Response Lenses

lttrsThingTypes :: Lens' ListThingTypesResponse [ThingTypeDefinition]

The thing types.

lttrsNextToken :: Lens' ListThingTypesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

lttrsResponseStatus :: Lens' ListThingTypesResponse Int

- | The response status code.