Network.AWS.IoT.ListThings

Description

Lists your things. Use the attributeName and attributeValue parameters to filter your things. For example, calling ListThings with attributeName=Color and attributeValue=Red retrieves all things in the registry that contain an attribute Color with the value Red .

This operation returns paginated results.

Creating a Request

listThings :: ListThings

Creates a value of ListThings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltAttributeValue - The attribute value used to search for things.

ltThingTypeName - The name of the thing type used to search for things.

ltNextToken - The token for the next set of results, or null if there are no additional results.

ltAttributeName - The attribute name used to search for things.

ltMaxResults - The maximum number of results to return in this operation.

data ListThings

The input for the ListThings operation.

See: listThings smart constructor.

Request Lenses

ltAttributeValue :: Lens' ListThings (Maybe Text)

The attribute value used to search for things.

ltThingTypeName :: Lens' ListThings (Maybe Text)

The name of the thing type used to search for things.

ltNextToken :: Lens' ListThings (Maybe Text)

The token for the next set of results, or null if there are no additional results.

ltAttributeName :: Lens' ListThings (Maybe Text)

The attribute name used to search for things.

ltMaxResults :: Lens' ListThings (Maybe Natural)

The maximum number of results to return in this operation.

Destructuring the Response

listThingsResponse

Arguments

:: Int
ltrsResponseStatus

-> ListThingsResponse

Creates a value of ListThingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltrsNextToken - The token for the next set of results, or null if there are no additional results.

ltrsThings - The things.

ltrsResponseStatus - -- | The response status code.

data ListThingsResponse

The output from the ListThings operation.

See: listThingsResponse smart constructor.

Response Lenses

ltrsNextToken :: Lens' ListThingsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

ltrsThings :: Lens' ListThingsResponse [ThingAttribute]

The things.

ltrsResponseStatus :: Lens' ListThingsResponse Int

- | The response status code.