Network.AWS.IoT.RegisterCACertificate

Description

Registers a CA certificate with AWS IoT. This CA certificate can then be used to sign device certificates, which can be then registered with AWS IoT. You can register up to 10 CA certificates per AWS account that have the same subject field. This enables you to have up to 10 certificate authorities sign your device certificates. If you have more than one CA certificate registered, make sure you pass the CA certificate when you register your device certificates with the RegisterCertificate API.

Creating a Request

registerCACertificate

Arguments

:: Text
rcacCaCertificate

-> Text
rcacVerificationCertificate

-> RegisterCACertificate

Creates a value of RegisterCACertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcacSetAsActive - A boolean value that specifies if the CA certificate is set to active.

rcacAllowAutoRegistration - Allows this CA certificate to be used for auto registration of device certificates.

rcacCaCertificate - The CA certificate.

rcacVerificationCertificate - The private key verification certificate.

data RegisterCACertificate

The input to the RegisterCACertificate operation.

See: registerCACertificate smart constructor.

Request Lenses

rcacSetAsActive :: Lens' RegisterCACertificate (Maybe Bool)

A boolean value that specifies if the CA certificate is set to active.

rcacAllowAutoRegistration :: Lens' RegisterCACertificate (Maybe Bool)

Allows this CA certificate to be used for auto registration of device certificates.

rcacCaCertificate :: Lens' RegisterCACertificate Text

The CA certificate.

rcacVerificationCertificate :: Lens' RegisterCACertificate Text

The private key verification certificate.

Destructuring the Response

registerCACertificateResponse

Arguments

:: Int
rcacrsResponseStatus

-> RegisterCACertificateResponse

Creates a value of RegisterCACertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcacrsCertificateARN - The CA certificate ARN.

rcacrsCertificateId - The CA certificate identifier.

rcacrsResponseStatus - -- | The response status code.

data RegisterCACertificateResponse

The output from the RegisterCACertificateResponse operation.

See: registerCACertificateResponse smart constructor.

Response Lenses

rcacrsCertificateARN :: Lens' RegisterCACertificateResponse (Maybe Text)

The CA certificate ARN.

rcacrsCertificateId :: Lens' RegisterCACertificateResponse (Maybe Text)

The CA certificate identifier.

rcacrsResponseStatus :: Lens' RegisterCACertificateResponse Int

- | The response status code.