Network.AWS.IoT.RejectCertificateTransfer

Description

Rejects a pending certificate transfer. After AWS IoT rejects a certificate transfer, the certificate status changes from PENDING_TRANSFER to INACTIVE .

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

This operation can only be called by the transfer destination. After it is called, the certificate will be returned to the source's account in the INACTIVE state.

Creating a Request

rejectCertificateTransfer

Arguments

:: Text
rctCertificateId

-> RejectCertificateTransfer

Creates a value of RejectCertificateTransfer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rctRejectReason - The reason the certificate transfer was rejected.

rctCertificateId - The ID of the certificate.

data RejectCertificateTransfer

The input for the RejectCertificateTransfer operation.

See: rejectCertificateTransfer smart constructor.

Request Lenses

rctRejectReason :: Lens' RejectCertificateTransfer (Maybe Text)

The reason the certificate transfer was rejected.

rctCertificateId :: Lens' RejectCertificateTransfer Text

The ID of the certificate.

Destructuring the Response

rejectCertificateTransferResponse :: RejectCertificateTransferResponse

Creates a value of RejectCertificateTransferResponse with the minimum fields required to make a request.

data RejectCertificateTransferResponse

See: rejectCertificateTransferResponse smart constructor.