Network.AWS.IoT.TransferCertificate

Description

Transfers the specified certificate to the specified AWS account.

You can cancel the transfer until it is acknowledged by the recipient.

No notification is sent to the transfer destination's account. It is up to the caller to notify the transfer target.

The certificate being transferred must not be in the ACTIVE state. You can use the UpdateCertificate API to deactivate it.

The certificate must not have any policies attached to it. You can use the DetachPrincipalPolicy API to detach them.

Creating a Request

transferCertificate

Arguments

:: Text
tcCertificateId

-> Text
tcTargetAWSAccount

-> TransferCertificate

Creates a value of TransferCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tcTransferMessage - The transfer message.

tcCertificateId - The ID of the certificate.

tcTargetAWSAccount - The AWS account.

data TransferCertificate

The input for the TransferCertificate operation.

See: transferCertificate smart constructor.

Request Lenses

tcTransferMessage :: Lens' TransferCertificate (Maybe Text)

The transfer message.

tcCertificateId :: Lens' TransferCertificate Text

The ID of the certificate.

tcTargetAWSAccount :: Lens' TransferCertificate Text

The AWS account.

Destructuring the Response

transferCertificateResponse

Arguments

:: Int
tcrsResponseStatus

-> TransferCertificateResponse

Creates a value of TransferCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tcrsTransferredCertificateARN - The ARN of the certificate.

tcrsResponseStatus - -- | The response status code.

data TransferCertificateResponse

The output from the TransferCertificate operation.

See: transferCertificateResponse smart constructor.

Response Lenses

tcrsTransferredCertificateARN :: Lens' TransferCertificateResponse (Maybe Text)

The ARN of the certificate.

tcrsResponseStatus :: Lens' TransferCertificateResponse Int

- | The response status code.