Network.AWS.IoT.UpdateCACertificate

Description

Updates a registered CA certificate.

Creating a Request

updateCACertificate

Arguments

:: Text
ucacCertificateId

-> UpdateCACertificate

Creates a value of UpdateCACertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucacNewStatus - The updated status of the CA certificate. Note: The status value REGISTER_INACTIVE is deprecated and should not be used.

ucacNewAutoRegistrationStatus - The new value for the auto registration status. Valid values are: ENABLE or DISABLE.

ucacCertificateId - The CA certificate identifier.

data UpdateCACertificate

The input to the UpdateCACertificate operation.

See: updateCACertificate smart constructor.

Request Lenses

ucacNewStatus :: Lens' UpdateCACertificate (Maybe CACertificateStatus)

The updated status of the CA certificate. Note: The status value REGISTER_INACTIVE is deprecated and should not be used.

ucacNewAutoRegistrationStatus :: Lens' UpdateCACertificate (Maybe AutoRegistrationStatus)

The new value for the auto registration status. Valid values are: ENABLE or DISABLE.

ucacCertificateId :: Lens' UpdateCACertificate Text

The CA certificate identifier.

Destructuring the Response

updateCACertificateResponse :: UpdateCACertificateResponse

Creates a value of UpdateCACertificateResponse with the minimum fields required to make a request.

data UpdateCACertificateResponse

See: updateCACertificateResponse smart constructor.